Refactoring iteration code. Added some improvements.
