+++ src/Infrastructure/Persistence/ApplicationContext.cs
@@ -1,18 +1,19 @@
-﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+﻿using Masny.QRAnimal.Infrastructure.Identity;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
-namespace Masny.QRAnimal.Infrastructure.Identity
+namespace Masny.QRAnimal.Infrastructure.Persistence
 {
     /// <summary>
-    /// Контекст для взаимодействия с пользователем.
+    /// Контекст для взаимодействия с приложения.
     /// </summary>
-    public class IdentityContext : IdentityDbContext<ApplicationUser>
+    public class ApplicationContext : IdentityDbContext<ApplicationUser>
     {
         /// <summary>
         /// Конструктор с параметрами.
         /// </summary>
         /// <param name="options">Параметры.</param>
-        public IdentityContext(DbContextOptions<IdentityContext> options)
+        public ApplicationContext(DbContextOptions<ApplicationContext> options)
             : base(options)
         {
             // Использовать, если не требуется выполнение миграции в режиме реального времени.
