+++ src/Web/appsettings.json
@@ -13,8 +13,8 @@
       {
         "Name": "MSSqlServer",
         "Args": {
-          "connectionString": "Server=(localdb)\\mssqllocaldb;Database=IdentityDatabase;Trusted_Connection=True;MultipleActiveResultSets=true",
-          "schemaName": "eventLogging",
+          "connectionString": "Server=(localdb)\\mssqllocaldb;Database=AnimalDatabase;Trusted_Connection=True;MultipleActiveResultSets=true",
+          "schemaName": "audit",
           "tableName": "Logs",
           "autoCreateSqlTable": true,
           "restrictedToMinimumLevel": "Information",
@@ -26,8 +26,7 @@
     "Enrich": [ "FromLogContext" ]
   },
   "ConnectionStrings": {
-    "AnimalConnection": "Server=(localdb)\\mssqllocaldb;Database=AnimalDatabase;Trusted_Connection=True;MultipleActiveResultSets=true",
-    "IdentityConnection": "Server=(localdb)\\mssqllocaldb;Database=IdentityDatabase;Trusted_Connection=True;MultipleActiveResultSets=true"
+    "ApplicationConnection": "Server=(localdb)\\mssqllocaldb;Database=AnimalDatabase;Trusted_Connection=True;MultipleActiveResultSets=true"
   },
   "AllowedHosts": "*"
 }
